+++ .eslintignore
@@ -1,0 +1,3 @@
+#Исключаем webpack
+webpack.config.js
+bundle.js
