+++ public/index.html
@@ -45,6 +45,6 @@
       <!-- Контент -->
     </main>
 
-    <script src="../src/main.js"></script>
+    <script type="module" src="./bundle.js"></script>
   </body>
 </html>
