+++ src/components/loadButton.js
@@ -1,0 +1,11 @@
+/**
+ * Создаем разметку Кнопка Load more»
+ * @return {void}
+ */
+const createLoadButtonTemplate = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
+
+export {createLoadButtonTemplate};
